🚑 fix(core): fix for critical race condition
This fixes an issue where a critical race condition occoured. The server would attempt to start a container with an invalid ID of a previously removed/stopped container. In attempt to still provide the service to the user, the server would then spin up a new container while failing to remove the old one. In this instance, two entries in the db for a user would be created. The new entries would conflict and the old ID (since it was added first) would be selected first. The server would then attempt to start more and more containers, yet still fail to provide service.